Util to flatten the gui_tree structure
Remove some duplication by providing an easy way to get a flattened
version of the gui_tree
