+++ backend/active_statistics/rq_tasks.py
@@ -9,11 +9,9 @@ from typing import Iterator, Optional, Type
 import plotly.graph_objects as go
 import sentry_sdk
 from active_statistics.app_logging import TASK, setup_task_logging
-from active_statistics.gui.gui import all_tabs
+from active_statistics.gui.gui import get_all_tabs
 from active_statistics.gui.image_tab import ImageTab
 from active_statistics.gui.plot_tabs import PlotTab
-from active_statistics.gui.tab_group import TabGroup
-from active_statistics.gui.tabs import Tab
 from active_statistics.gui.trivia_tabs import TableTab
 from active_statistics.strava_custom_rate_limiters import DetailedTaskRateLimiter
 from active_statistics.utils import redis
@@ -169,17 +167,7 @@ def process_activities(athlete_id: int, detailed: bool) -> None:
     detailed activities to make plots. This gives us a bunch of data. This data
     is then saved to s3.
     """
-
-    def flatten(tabs: list[Tab | TabGroup]) -> list[Tab]:
-        flattened_tabs = []
-        for tab in tabs:
-            if isinstance(tab, TabGroup):
-                flattened_tabs.extend(flatten(tab.children))
-            else:
-                flattened_tabs.append(tab)
-        return flattened_tabs
-
-    for tab in flatten(get_all_tabs()):
+    for tab in get_all_tabs():
         if tab.is_detailed() == detailed:
             try:
                 activity_iterator: Iterator[Activity] = (
@@ -248,10 +236,3 @@ def log(log_fuction, message: str) -> None:
     if isinstance(job, Job):
         job.meta["message"] = message
         job.save()
-
-
-def get_all_tabs() -> list[Tab | TabGroup]:
-    """
-    Returns all tabs. Exists as it's own function so I can mock it in tests.
-    """
-    return all_tabs
