+++ backend/active_statistics/server.py
@@ -10,7 +10,7 @@ from active_statistics.communication_schema import (
     DataStatusMessage,
     RefreshStatusMessage,
 )
-from active_statistics.gui.gui import all_tabs
+from active_statistics.gui.gui import get_all_tabs, tab_tree
 from active_statistics.gui.tab_group import TabGroup
 from active_statistics.gui.tabs import Tab
 from active_statistics.utils import human_readable_time, redis, rq
@@ -448,24 +448,12 @@ def tabs_route() -> Response:
             name=tab.name, key=tab.get_key(), type=tab.get_type(), items=items
         ).to_dict()
 
-    return make_response(jsonify(expand_tabs(all_tabs)))
+    return make_response(jsonify(expand_tabs(tab_tree)))
 
 
 # Before the app starts, we want to generate all the routes for our tabs.
-for tab in all_tabs:
-
-    def generate_and_register_routes_for_children(tab_group: TabGroup) -> None:
-        for tab in tab_group.children:
-            if isinstance(tab, Tab):
-                tab.generate_and_register_route(app, evm)
-            if isinstance(tab, TabGroup):
-                generate_and_register_routes_for_children(tab)
-
-    if isinstance(tab, Tab):
-        tab.generate_and_register_route(app, evm)
-    if isinstance(tab, TabGroup):
-        generate_and_register_routes_for_children(tab)
-
+for tab in get_all_tabs():
+    tab.generate_and_register_route(app, evm)
 
 if not evm.is_production():
     app.run(debug=True)
